Changed to include integration tests currently only with mocked jira
