+++ ReleaseProcessAutomation.IntegrationTests/ReleaseFromSupportTests.cs
@@ -29,9 +29,8 @@ internal class ReleaseFromSupportTests : IntegrationTestSetup
 
     Program.Console = TestConsole;
 
-    Program.Main(new[] { "Release-Version" });
-    var output = TestConsole.Output.Replace("\n", "");
-
-    Assert.That(output, Does.Contain("You have to be on either a 'hotfix/*' or 'release/*' or 'develop' or 'master' branch to release a version"));
+    Assert.That(() => RunProgram(new[] { "Release-Version" }), 
+        Throws.InstanceOf<InvalidOperationException>().
+            With.Message.EqualTo("You have to be on either a 'hotfix/*' or 'release/*' or 'develop' or 'master' branch to release a version"));
   }
 }
